tests/output/items_api/pyproject.toml
[tool.poetry]
name = "Items Api"
version = "0.1.0"
description = "API Generated by Median Code"
authors = ["Median Code"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"


[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"